+++ src/controls/folderPicker/FolderPicker.tsx
@@ -48,7 +48,7 @@ export class FolderPicker extends React.Component<IFolderPickerProps, IFolderPic
             }
             {this.state.selectedFolder &&
               <div className={styles.selectFolder}>
-                <Link className={styles.selectFolder} target='_blank' data-interception="off" id={this._folderLinkId} href={this.state.selectedFolder.ServerRelativeUrl}>
+                <Link className={styles.selectedLink} target='_blank' data-interception="off" id={this._folderLinkId} href={this.state.selectedFolder.ServerRelativeUrl}>
                   <span title={this.state.selectedFolder.Name}>{this.state.selectedFolder.Name}</span>
                 </Link>
                 <IconButton